fix(particle-editor): fix compilation errors — API compatibility fixes
- GlStateManager: use com.mojang.blaze3d.platform (not .systems)
- blendFuncSeparate: use SourceFactor/DestFactor pairing
- Tint.parseGradient: rename local variable to avoid shadowing field
- ParticleCurve: cast double to float in bezier chain compute
- Billboard: use Matrix4f.set(16 floats) for emitter transform facing
- UIParticleSchemeAppearanceSection: use IKey.raw/constant instead
  of non-existent IKey.literal/str, remove non-existent decimalSingle()

+++ src/client/java/mchorse/bbs_mod/particles/ParticleCurve.java
@@ -140,7 +140,7 @@ public class ParticleCurve
 
                 float t0 = prev.getKey();
                 float t1 = entry.getKey();
-                float localT = (factor - t0) / (t1 - t0);
+                float localT = (float) ((factor - t0) / (t1 - t0));
 
                 BezierChainNode n0 = prev.getValue();
                 BezierChainNode n1 = entry.getValue();
